refactor(ai-service): extract transformer trainer to pipelines/training/
Completes the transformer file split started in PR #4. Moves the
training orchestrator out of app/ml/transformer/model.py — model.py
is now actually just model code (1012 LOC, down from 2737 pre-PR4).

Moved (884 LOC):
  app/ml/transformer/model.py
    ├─ ExpertActionTrainer       (the epoch loop + checkpoint mgmt)
    └─ prepare_and_train_coach_transformer_model  (async entry)
                                 → app/pipelines/training/transformer_trainer.py

Importer updated:
  app/pipelines/training/full_dataset.py — was importing the entry
  from app.ml.transformer.model; now imports from sibling
  transformer_trainer.

model.py now contains exactly what its name suggests:
  - _safe_number, make_json_safe (JSON helpers)
  - PositionalEncoding (model layer)
  - ExpertActionTransformer (the actual nn.Module)

Pre-PR4: 2737 LOC, 8 top-level constructs across 4 bands
Post-PR5: 1012 LOC, 4 top-level constructs, all model code

Validated:
  • ExpertActionTrainer + prepare_and_train_coach_transformer_model
    importable from new path; asyncio.iscoroutinefunction(entry) == True
  • Model classes still importable from app.ml.transformer.model
  • lint-imports 11 kept, 0 broken
  • pytest 24 passed
  • /health 200
